address bypassing out-of-index exception in List.take/skip,
The out-of-index exception is due to the List.take/skip functions
from the base class library. The current used List.take/skip functions
are from FSharpx.Collections, which don't throw if the take(or skip) number
is great than the list length.

This issue can be addressed using the min value between the take(or skip) number
and the list length.